Updated build script

+++ build.sh
@@ -10,7 +10,11 @@ g++ -shared -fPIC -o libNativeLib.so -I$2/include -I$2/include/linux wordrain.cp
 java -Djava.library.path=. Main
 
 # remove .class files
-rm *.class
-rm Main_WordRain.h
-rm message.txt
-rm libNativeLib.so
+echo "Do you want to remove machine-generated files?"
+read a
+if [[ "$a" == "y" ]] ; then
+    rm *.class
+    rm Main_WordRain.h
+    rm message.txt
+    rm libNativeLib.so
+fi
